changes after team comments

+++ Frontend/src/components/input_formats/Ranking_value_exp.vue
@@ -110,7 +110,10 @@ export default {
         rowUnHovered(item){
             this.$refs.map.changeOpacity(item.item_name,0.6);
         },
-        submit(){
+        async submit(){
+            if (!confirm("Once you press OK you can't go back and change your choices"))
+                return;
+                
             let time=new Date().getTime();
             localStorage.setItem('budgeting_finish',JSON.stringify(time));
             let final_items=[];
@@ -121,6 +124,8 @@ export default {
             });
             console.log(final_items);
             localStorage.setItem('final_items',JSON.stringify(final_items));
+
+            await this.$parent.addExperiment();
             this.$router.push("/Consistency");
 
         }
@@ -164,6 +169,7 @@ export default {
             float: left;
             width: 100%;
             padding: 10px;
+            margin-left: 400px;
         }
         .apexchart{
             position:relative;
